refactor(metrics): clean up VLM benchmark helpers and update dependencies
- Removed the `vlm_benchmark_helpers.py` file as it was no longer needed.
- Updated `pyproject.toml` to remove the `mine-replicate` dependency.
- Refined documentation and code comments for clarity in VLM metrics handling.
- Adjusted test cases to reflect the removal of the benchmark helpers.
