+++ tests/evaluation/test_oneig_metrics.py
@@ -1,4 +1,4 @@
-"""Tests for OneIG alignment dependency masking and metric wiring."""
+"""Tests for OneIG alignment (masking, wiring) and OneIG reasoning (LLM2CLIP)."""
 
 from unittest.mock import MagicMock
 
@@ -12,6 +12,11 @@ from pruna.evaluation.metrics.metric_oneig_alignment import (
     aggregate_oneig_alignment_per_cell,
     apply_oneig_dependency_mask,
 )
+from pruna.evaluation.metrics.metric_oneig_reasoning import (
+    OneIGReasoningMetric,
+    _LLM2CLIPScorer,
+)
+from pruna.evaluation.metrics.registry import MetricRegistry
 from pruna.evaluation.metrics.vlm_base import BaseVLM
 
 
@@ -46,15 +51,11 @@ def test_apply_oneig_dependency_mask_uses_raw_parent_not_filtered_for_chain() ->
 
 def test_apply_oneig_dependency_mask_grandchild_chain() -> None:
     """3-level chain: grandparent No masks parent; grandchild uses raw parent (stays 1.0)."""
-    # q1=grandparent(No), q2=parent(Yes) depends on q1, q3=grandchild(Yes) depends on q2
     raw_scores = {1: 0.0, 2: 1.0, 3: 1.0}
     dependencies = {2: [1], 3: [2]}
     filtered = apply_oneig_dependency_mask(raw_scores, dependencies)
-    # q2 masked because q1 is No (raw[1]=0.0)
     assert filtered[2] == 0.0
-    # q3 uses raw[2]=1.0, NOT filtered[2]=0.0 → stays 1.0
     assert filtered[3] == 1.0
-    # q1 unchanged
     assert filtered[1] == 0.0
 
 
@@ -136,3 +137,92 @@ def test_to_oneig_record_strips_null_questions_and_dependencies() -> None:
     assert "21" not in record["dependencies"]
     assert record["questions"] == {"1": "Is there a cat?"}
     assert record["dependencies"] == {"1": [0]}
+
+
+def _make_mock_scorer(return_value: float = 0.5) -> MagicMock:
+    mock = MagicMock(spec=_LLM2CLIPScorer)
+    mock.score.return_value = [return_value]
+    return mock
+
+
+@pytest.mark.cpu
+def test_oneig_reasoning_uses_gt_answer_from_aux() -> None:
+    """Metric reads reasoning_gt_answer from aux."""
+    mock_scorer = _make_mock_scorer(0.7)
+    metric = OneIGReasoningMetric(scorer=mock_scorer, device="cpu")
+    images = torch.rand(1, 3, 64, 64)
+    aux = {"reasoning_gt_answer": "A blue circle"}
+    metric.update(["p"], [aux], images)
+    result = metric.compute()
+    assert result.name == "oneig_reasoning"
+    assert result.result == 0.7
+    mock_scorer.score.assert_called_once()
+    call_args = mock_scorer.score.call_args
+    assert call_args[0][1] == "A blue circle"
+
+
+@pytest.mark.cpu
+def test_oneig_reasoning_averages_per_sample_scores() -> None:
+    """Compute returns mean of per-sample scores."""
+    mock_scorer = _make_mock_scorer(0.5)
+    metric = OneIGReasoningMetric(scorer=mock_scorer, device="cpu")
+    images = torch.rand(2, 3, 64, 64)
+    aux_list = [
+        {"reasoning_gt_answer": "First answer"},
+        {"reasoning_gt_answer": "Second answer"},
+    ]
+    metric.update(["p1", "p2"], aux_list, images)
+    result = metric.compute()
+    assert result.result == 0.5
+    assert mock_scorer.score.call_count == 2
+
+
+@pytest.mark.cpu
+def test_oneig_reasoning_missing_gt_raises() -> None:
+    """Missing GT answer raises ValueError."""
+    mock_scorer = _make_mock_scorer(0.8)
+    metric = OneIGReasoningMetric(scorer=mock_scorer, device="cpu")
+    images = torch.rand(1, 3, 64, 64)
+    aux = {}
+    with pytest.raises(ValueError, match="reasoning_gt_answer"):
+        metric.update(["p"], [aux], images)
+    mock_scorer.score.assert_not_called()
+
+
+@pytest.mark.cpu
+def test_oneig_reasoning_scorer_none_raises() -> None:
+    """When scorer returns None, metric raises RuntimeError."""
+    mock_scorer = _make_mock_scorer()
+    mock_scorer.score.return_value = None
+    metric = OneIGReasoningMetric(scorer=mock_scorer, device="cpu")
+    images = torch.rand(1, 3, 64, 64)
+    aux = {"reasoning_gt_answer": "Some answer"}
+    with pytest.raises(RuntimeError, match="no scores"):
+        metric.update(["p"], [aux], images)
+
+
+@pytest.mark.cpu
+def test_oneig_reasoning_compute_without_update_raises() -> None:
+    """Compute with no updates raises RuntimeError."""
+    mock_scorer = _make_mock_scorer()
+    metric = OneIGReasoningMetric(scorer=mock_scorer, device="cpu")
+    with pytest.raises(RuntimeError, match="no samples were scored"):
+        metric.compute()
+
+
+@pytest.mark.cpu
+def test_oneig_reasoning_has_metric_registered() -> None:
+    """oneig_reasoning is available via MetricRegistry (lazy)."""
+    assert MetricRegistry.has_metric("oneig_reasoning")
+
+
+@pytest.mark.cpu
+def test_transformers_major_version_supported_for_oneig_reasoning() -> None:
+    """Enforce pyproject ``transformers<5`` expectation for LLM2CLIP loading."""
+    import transformers
+
+    major = int(transformers.__version__.split(".", 1)[0])
+    assert major < 5, (
+        "oneig_reasoning expects transformers 4.x (see pyproject.toml); "
+        "5.x from_pretrained buffer initialization can break CLIP/Llama stacks."
+    )
